// packages/api/src/vfs/tools/toolEnum.ts
export enum VFSTool {
  read_file = 'read_file',
  list_directory = 'list_directory',
  find_files = 'find_files',
  search_text = 'search_text',
  get_file_metadata = 'get_file_metadata',
  get_file_tree = 'get_file_tree',
  count_lines = 'count_lines',
  search_symbol = 'search_symbol',
  create_file = 'create_file',
  edit_file = 'edit_file',
  delete_file = 'delete_file',
  rename_file = 'rename_file',
}